+++ spec/interjectable_spec.rb
@@ -6,12 +6,17 @@ describe Interjectable do
   shared_examples_for "an interjectable class" do
     let(:instance) { klass.new }
 
-    def expect_injected_methods(injecting_class, class_instance, *getter_methods)
+    def expect_injected_methods(injecting_class, class_instance, include_super, *getter_methods)
+      expect(injecting_class.injected_methods(include_super))
+        .to match_array(class_instance.injected_methods(include_super))
+
       setter_methods = getter_methods.map { |m| :"#{m}=" }
-      expect(injecting_class.injected_methods)
-        .to match_array(getter_methods + setter_methods + [Interjectable::INJECTED_METHODS])
-      expect(injecting_class.injected_methods)
-        .to match_array(class_instance.injected_methods)
+      expected_methods = getter_methods + setter_methods + [:injected_methods]
+      if include_super && injecting_class.superclass.respond_to?(:injected_methods)
+        expected_methods.append(*injecting_class.superclass.injected_methods(include_super)).uniq!
+      end
+
+      expect(injecting_class.injected_methods(include_super)).to match_array(expected_methods)
     end
 
     describe "#inject" do
@@ -27,7 +32,7 @@ describe Interjectable do
         instance.some_dependency = 'aaa'
         expect(instance.some_dependency).to eq('aaa')
 
-        expect_injected_methods(klass, instance, :some_dependency)
+        expect_injected_methods(klass, instance, true, :some_dependency)
       end
 
       it "lazy-loads the default block" do
@@ -35,7 +40,7 @@ describe Interjectable do
         expect(instance.some_dependency).to eq(:service)
         expect(instance.instance_variable_get("@some_dependency")).not_to be_nil
 
-        expect_injected_methods(klass, instance, :some_dependency)
+        expect_injected_methods(klass, instance, true, :some_dependency)
       end
 
       it "allows transitive dependencies (via instance_eval)" do
@@ -43,7 +48,7 @@ describe Interjectable do
         klass.inject(:second_dependency) { :value }
 
         expect(instance.first_dependency).to eq(:value)
-        expect_injected_methods(klass, instance, :some_dependency, :first_dependency, :second_dependency)
+        expect_injected_methods(klass, instance, true, :some_dependency, :first_dependency, :second_dependency)
       end
 
       it "calls dependency block once, even with a falsy value" do
@@ -53,7 +58,7 @@ describe Interjectable do
         2.times { expect(instance.some_falsy_dependency).to be_nil }
         expect(count).to eq(1)
 
-        expect_injected_methods(klass, instance, :some_dependency, :some_falsy_dependency)
+        expect_injected_methods(klass, instance, true, :some_dependency, :some_falsy_dependency)
       end
 
       it "errors when injecting the same dependency multiple times" do
@@ -76,7 +81,7 @@ describe Interjectable do
         instance.foo = 2
         expect(instance.good_dependency).to eq(2)
 
-        expect_injected_methods(klass, instance, :some_dependency, :good_dependency)
+        expect_injected_methods(klass, instance, true, :some_dependency, :good_dependency)
       end
 
       context "with a dependency on another class" do
@@ -90,7 +95,7 @@ describe Interjectable do
           instance.some_other_class = :fake_other_class
 
           expect(instance.some_other_class).to eq(:fake_other_class)
-          expect_injected_methods(klass, instance, :some_dependency, :some_other_class)
+          expect_injected_methods(klass, instance, true, :some_dependency, :some_other_class)
         end
       end
 
@@ -102,8 +107,8 @@ describe Interjectable do
           subclass.inject(:some_dependency) { :some_other_value }
           expect(subclass_instance.some_dependency).to eq(:some_other_value)
 
-          expect_injected_methods(klass, instance, :some_dependency)
-          expect_injected_methods(subclass, subclass_instance, :some_dependency)
+          expect_injected_methods(klass, instance, true, :some_dependency)
+          expect_injected_methods(subclass, subclass_instance, true, :some_dependency)
         end
 
         it "allows injection on the subclass without injecting on the superclass" do
@@ -111,8 +116,31 @@ describe Interjectable do
           expect(subclass_instance.subclass_dependency).to eq(:brand_new_value)
           expect { instance.subclass_dependency }.to raise_error(NoMethodError)
 
-          expect_injected_methods(klass, instance, :some_dependency)
-          expect_injected_methods(subclass, subclass_instance, :some_dependency, :subclass_dependency)
+          expect_injected_methods(klass, instance, true, :some_dependency)
+          expect_injected_methods(klass, instance, false, :some_dependency)
+
+          expect_injected_methods(subclass, subclass_instance, true, :some_dependency, :subclass_dependency)
+          expect_injected_methods(subclass, subclass_instance, false, :subclass_dependency)
+        end
+
+        context "with a chain of subclasses" do
+          let(:lower_subclass) { Class.new(subclass) }
+          let(:lower_subclass_instance) { lower_subclass.new }
+
+          it "retrieves injected methods from all ancestors when requested" do
+            subclass.inject(:subclass_dependency) { :subclass_value }
+            lower_subclass.inject(:lower_subclass_dependency) { :lower_subclass_value }
+
+            expect_injected_methods(klass, instance, true, :some_dependency)
+            expect_injected_methods(klass, instance, false, :some_dependency)
+
+            expect_injected_methods(subclass, subclass_instance, true, :some_dependency, :subclass_dependency)
+            expect_injected_methods(subclass, subclass_instance, false, :subclass_dependency)
+
+            expect_injected_methods(lower_subclass, lower_subclass_instance, true,
+                                    :some_dependency, :subclass_dependency, :lower_subclass_dependency)
+            expect_injected_methods(lower_subclass, lower_subclass_instance, false, :lower_subclass_dependency)
+          end
         end
       end
     end
@@ -132,16 +160,16 @@ describe Interjectable do
         instance.static_dependency = 'aaa'
         expect(instance.static_dependency).to eq('aaa')
 
-        expect_injected_methods(klass, instance, :static_dependency)
-        expect_injected_methods(klass, other_instance, :static_dependency)
+        expect_injected_methods(klass, instance, true, :static_dependency)
+        expect_injected_methods(klass, other_instance, true, :static_dependency)
       end
 
       it "adds a class method and setter" do
         klass.static_dependency = 'aaa'
         expect(klass.static_dependency).to eq('aaa')
 
-        expect_injected_methods(klass, instance, :static_dependency)
-        expect_injected_methods(klass, other_instance, :static_dependency)
+        expect_injected_methods(klass, instance, true, :static_dependency)
+        expect_injected_methods(klass, other_instance, true, :static_dependency)
       end
 
       it "shares a value across all instances of a class" do
@@ -149,8 +177,8 @@ describe Interjectable do
         expect(other_instance.static_dependency).to eq('bbb')
         expect(klass.static_dependency).to eq('bbb')
 
-        expect_injected_methods(klass, instance, :static_dependency)
-        expect_injected_methods(klass, other_instance, :static_dependency)
+        expect_injected_methods(klass, instance, true, :static_dependency)
+        expect_injected_methods(klass, other_instance, true, :static_dependency)
       end
 
       it "calls its dependency block once across all instances" do
@@ -163,8 +191,8 @@ describe Interjectable do
 
         expect(count).to eq(1)
 
-        expect_injected_methods(klass, instance, :static_dependency, :falsy_static_dependency)
-        expect_injected_methods(klass, other_instance, :static_dependency, :falsy_static_dependency)
+        expect_injected_methods(klass, instance, true, :static_dependency, :falsy_static_dependency)
+        expect_injected_methods(klass, other_instance, true, :static_dependency, :falsy_static_dependency)
       end
 
       it "errors when inject_static-ing a dependency multiple times" do
@@ -206,9 +234,9 @@ describe Interjectable do
           expect(subclass_instance.static_dependency).to eq('ccc')
           expect(subclass.static_dependency).to eq('ccc')
 
-          expect_injected_methods(klass, instance, :static_dependency)
-          expect_injected_methods(klass, other_instance, :static_dependency)
-          expect_injected_methods(subclass, subclass_instance, :static_dependency)
+          expect_injected_methods(klass, instance, true, :static_dependency)
+          expect_injected_methods(klass, other_instance, true, :static_dependency)
+          expect_injected_methods(subclass, subclass_instance, true, :static_dependency)
         end
 
         it "does not error if the method exists on the super klass" do
@@ -216,9 +244,9 @@ describe Interjectable do
           expect(subclass_instance.static_dependency).to eq(:some_other_value)
           expect(subclass.static_dependency).to eq(:some_other_value)
 
-          expect_injected_methods(klass, instance, :static_dependency)
-          expect_injected_methods(klass, other_instance, :static_dependency)
-          expect_injected_methods(subclass, subclass_instance, :static_dependency)
+          expect_injected_methods(klass, instance, true, :static_dependency)
+          expect_injected_methods(klass, other_instance, true, :static_dependency)
+          expect_injected_methods(subclass, subclass_instance, true, :static_dependency)
         end
 
         it "does not error when lazily setting the dep from a subclass first" do
@@ -226,18 +254,18 @@ describe Interjectable do
           expect(klass.static_dependency).to eq(:some_value)
           expect(subclass.static_dependency).to eq(:some_value)
 
-          expect_injected_methods(klass, instance, :static_dependency)
-          expect_injected_methods(klass, other_instance, :static_dependency)
-          expect_injected_methods(subclass, subclass_instance, :static_dependency)
+          expect_injected_methods(klass, instance, true, :static_dependency)
+          expect_injected_methods(klass, other_instance, true, :static_dependency)
+          expect_injected_methods(subclass, subclass_instance, true, :static_dependency)
         end
 
         it "only defines the class variable on the injecting class" do
           expect(subclass.static_dependency).to eq(:some_value)
           expect(klass.class_variable_get(:@@static_dependency)).to eq(:some_value)
 
-          expect_injected_methods(klass, instance, :static_dependency)
-          expect_injected_methods(klass, other_instance, :static_dependency)
-          expect_injected_methods(subclass, subclass_instance, :static_dependency)
+          expect_injected_methods(klass, instance, true, :static_dependency)
+          expect_injected_methods(klass, other_instance, true, :static_dependency)
+          expect_injected_methods(subclass, subclass_instance, true, :static_dependency)
         end
 
         it "allows injection on the subclass without injecting on the superclass" do
@@ -246,8 +274,31 @@ describe Interjectable do
           expect { klass.static_subclass_dependency }.to raise_error(NoMethodError)
           expect { instance.static_subclass_dependency }.to raise_error(NoMethodError)
 
-          expect_injected_methods(klass, instance, :static_dependency)
-          expect_injected_methods(subclass, subclass_instance, :static_dependency, :static_subclass_dependency)
+          expect_injected_methods(klass, instance, true, :static_dependency)
+          expect_injected_methods(klass, instance, false, :static_dependency)
+
+          expect_injected_methods(subclass, subclass_instance, true, :static_dependency, :static_subclass_dependency)
+          expect_injected_methods(subclass, subclass_instance, false, :static_subclass_dependency)
+        end
+
+        context "with a chain of subclasses" do
+          let(:lower_subclass) { Class.new(subclass) }
+          let(:lower_subclass_instance) { lower_subclass.new }
+
+          it "retrieves injected methods from all ancestors when requested" do
+            subclass.inject_static(:static_subclass_dependency) { :subclass_value }
+            lower_subclass.inject_static(:static_lower_subclass_dependency) { :lower_subclass_value }
+
+            expect_injected_methods(klass, instance, true, :static_dependency)
+            expect_injected_methods(klass, instance, false, :static_dependency)
+
+            expect_injected_methods(subclass, subclass_instance, true, :static_dependency, :static_subclass_dependency)
+            expect_injected_methods(subclass, subclass_instance, false, :static_subclass_dependency)
+
+            expect_injected_methods(lower_subclass, lower_subclass_instance, true,
+                                    :static_dependency, :static_subclass_dependency, :static_lower_subclass_dependency)
+            expect_injected_methods(lower_subclass, lower_subclass_instance, false, :static_lower_subclass_dependency)
+          end
         end
       end
     end